requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==4.2.11
django-browser-reload==1.12.1
django-tailwind==3.6.0
djangorestframework==3.14.0
dlib @ file:///K:/app/biometric/biometric-app/dlib-19.22.99-cp39-cp39-win_amd64.whl#sha256=0fd828405c77af2df1ff6a09964a8ba7f14838c538d9017046804f353e9b4bc2
face-recognition==1.3.0
face_recognition_models==0.3.0
idna==3.6
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.2.0
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
tailwind==3.1.5b0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
